Clean up the breakpoint handler, and push through some arch abstractions.
This commit pushes breakpoint codes to the architecture definitions, and
cleans up the handler routine.  It also reduces the number of operations
required when installing and uninstalling breakpoints.

The text_base parameter has been removed from the bina_trace_init
routine, and /temporarily/ hardcoded into the start_child routine until
a reliable ptrace method is discovered for determining this.