Don't bother using ptrace to read original memory, just read it directly.
This is relatively safe because we're reading directly from the original file
and not using ptrace to inspect memory.  It doesn't take into account memory
that's been dynamically modified at run-time from the original binary, but
that scenario isn't likely (if at all possible in the text segment).

It also guards against installing a breakpoint in the same place.  Although
we'll still go through the installation dance, it won't corrupt memory by
installing breakpoint code over the top of breakpoint code, and losing the
original code.