Introduce the bina-test application.
The bina-test application is an application that tests various parts of the
bina API.  It's simply a playground for the test library.  Currently, it
instruments an application with breakpoints at each basic block, and prints
out the block indicies as they are entered.

It also generates a DOT graph with the basic block linkages.